refactored code structure for App.js and index.js
