+++ .gitignore
@@ -11,7 +11,7 @@ server/node_modules
 /coverage
 
 # production
-/build
+*/build
 
 # misc
 .DS_Store
