+++ client/src/index.js
@@ -1,10 +1,20 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import {
+  createTheme,
+  responsiveFontSizes,
+  ThemeProvider,
+} from '@material-ui/core/styles';
 import App from './app/App.js';
+
+let theme = createTheme();
+theme = responsiveFontSizes(theme);
 
 ReactDOM.render(
   <React.StrictMode>
-    <App />
+    <ThemeProvider theme={theme}>
+      <App />
+    </ThemeProvider>
   </React.StrictMode>,
   document.getElementById('root')
 );
